agent/requirements.txt
fastapi>=0.127.1
uvicorn>=0.40.0
python-dotenv>=1.2.1